# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Legend Bucks rewards platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          description: Unauthenticated
  /auth/login:
    post:
      operationId: requestMagicLink
      tags: [auth]
      summary: Request a magic link login email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MagicLinkRequest"
      responses:
        "200":
          description: Magic link sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MagicLinkResponse"
        "400":
          description: Invalid email
  /auth/verify:
    post:
      operationId: verifyMagicLink
      tags: [auth]
      summary: Verify a magic link token and create session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MagicLinkVerify"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "400":
          description: Invalid or expired token
...
  /auth/invite:
    post:
      operationId: inviteEmployee
      tags: [auth]
      summary: Admin invites a new employee by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteInput"
      responses:
        "201":
          description: Invitation sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "400":
          description: Bad request
        "403":
          description: Forbidden

  # Employees
...
  /employees/{id}/balance:
    get:
      operationId: getEmployeeBalance
      tags: [employees]
      summary: Get an employee's current balance calculated from ledger
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceSummary"
...
  /budgets:
    get:
      operationId: listBudgets
      tags: [budgets]
      summary: List monthly budgets
      parameters:
        - name: managerId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: string
            description: "Format: YYYY-MM"
      responses:
        "200":
          description: List of budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Budget"
    post:
      operationId: assignBudget
      tags: [budgets]
      summary: Assign or update monthly budget for a manager
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetInput"
      responses:
        "201":
          description: Budget assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
        "400":
          description: Bad request
  /budgets/{id}:
    get:
      operationId: getBudget
      tags: [budgets]
      summary: Get a budget record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
        "404":
          description: Not found
...
components:
  schemas:
...
    CurrentUser:
      type: object
      required: [id, firstName, lastName, email, role, status]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, manager, team_member]
        department:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        managerId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, inactive, invited]
        balance:
          type: integer
          description: Current Legend Bucks balance
    Employee:
      type: object
      required: [id, firstName, lastName, email, role, status, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        department:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        managerId:
          type: ["integer", "null"]
        managerName:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, manager, team_member]
        status:
          type: string
          enum: [active, inactive, invited]
        balance:
          type: ["integer", "null"]
        createdAt:
          type: string
...
    BalanceSummary:
      type: object
      required: [employeeId, balance, totalReceived, totalSpent]
      properties:
        employeeId:
          type: integer
        balance:
          type: integer
        totalReceived:
          type: integer
        totalSpent:
          type: integer
    MagicLinkRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    MagicLinkResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
        token:
          type: ["string", "null"]
          description: Only included in dev mode for easy testing
    MagicLinkVerify:
      type: object
      required: [token]
      properties:
        token:
          type: string
    InviteInput:
      type: object
      required: [email, firstName, lastName, role]
      properties:
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [admin, manager, team_member]
        department:
          type: string
        location:
          type: string
        managerId:
          type: ["integer", "null"]
    Budget:
      type: object
      required: [id, managerId, month, totalAmount, usedAmount, createdAt]
      properties:
        id:
          type: integer
        managerId:
          type: integer
        managerName:
          type: string
        month:
          type: string
          description: "Format: YYYY-MM"
        totalAmount:
          type: integer
        usedAmount:
          type: integer
        remainingAmount:
          type: integer
        createdAt:
          type: string
    BudgetInput:
      type: object
      required: [managerId, month, totalAmount]
      properties:
        managerId:
          type: integer
        month:
          type: string
          description: "Format: YYYY-MM"
        totalAmount:
          type: integer
          minimum: 0
...