# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Legend Bucks rewards platform API
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeRole } from './employeeRole';
import type { EmployeeStatus } from './employeeStatus';

export interface Employee {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  managerId?: number | null;
  /** @nullable */
  managerName?: string | null;
  role: EmployeeRole;
  status: EmployeeStatus;
  /** @nullable */
  balance?: number | null;
  createdAt: string;
}